requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pgvector==0.2.4

# OpenAI
openai==1.10.0

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
langchain-text-splitters==0.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3

# CORS
python-jose[cryptography]==3.3.0
